requirements.txt
captum==0.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.25.0
contourpy==1.1.0
cycler==0.11.0
filelock==3.9.0
fonttools==4.41.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
lit==15.0.7
MarkupSafe==2.1.2
matplotlib==3.7.2
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
packaging==23.1
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
sympy==1.11.1
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.4.0
urllib3==1.26.13
